libchrome: Uprev the library to r395517 from Chromium
Pulled the latest and greatest version of libchrome from Chromium.

The merge was done against r395517 which corresponds to git commit
ebdcb576bb346af95b8ad219f6250daf63122f98 of May 23, 2016

Notable changes are:
- scoped_ptr was removed in favor of std::unique_ptr
- base/thread_task_runner_handle.h was moved to base/threading.

BUG: 28985443
TEST: All tests in libchrome_test pass on dragonboard-eng build

+++ base/hash.cc
@@ -8,7 +8,7 @@
 
 namespace base {
 
-uint32_t SuperFastHash(const char* data, int len) {
+uint32_t SuperFastHash(const char* data, size_t len) {
   std::hash<std::string> hash_fn;
   return hash_fn(std::string(data, len));
 }
